7.5
Lots of stuff done
found a way of saving css in localstorage, not compleatly done, need to
make it easy to save and have not tested if its possible to communicate
to the bg page yet.
rewrote the settings html page since chrome does not allow for inline
execution of functions, need to add handlers througth a script.